+++ toolchain/scripts/modules/start/framer.ps1
@@ -1,5 +1,3 @@
-Import-Module "$($config.project.neocorePath)\toolchain\scripts\modules\install\sdk.ps1"
-
 function Start-Framer {
     $exe = "$($Config.project.buildPath)\neodev-sdk\m68k\bin\Framer.exe"