+++ toolchain/scripts/modules/start/mame.ps1
@@ -1,8 +1,7 @@
-Import-Module "$($Config.project.neocorePath)\toolchain\scripts\modules\write\mame.ps1"
-
 function Start-Mame {
   $exeName = [System.IO.Path]::GetFileName($Config.project.emulator.mame.exeFile)
   $mamePath = Split-Path $Config.project.emulator.mame.exeFile
+  $mamePath = Get-TemplatePath -Path $mamePath
 
   # TODO : remove at v3